blueprint split of job related pages
move job related pages into the jobs flask blueprint
