+++ arm/ui/__init__.py
@@ -64,11 +64,13 @@ from arm.ui.logs.logs import route_logs  # noqa: E402,F811
 from arm.ui.auth.auth import route_auth  # noqa: E402,F811
 from arm.ui.database.database import route_database  # noqa: E402,F811
 from arm.ui.history.history import route_history  # noqa: E402,F811
+from arm.ui.jobs.jobs import route_jobs  # noqa: E402,F811
 app.register_blueprint(route_settings)
 app.register_blueprint(route_logs)
 app.register_blueprint(route_auth)
 app.register_blueprint(route_database)
 app.register_blueprint(route_history)
+app.register_blueprint(route_jobs)
 
 # Remove GET/page loads from logging
 import logging  # noqa: E402,F811
