+++ arm/ui/jobs/__init__.py
@@ -1,0 +1,3 @@
+"""
+ARM route blueprint for jobs pages
+"""
