+++ arm/ui/jobs/templates/jobdetail.html
