+++ arm/ui/routes.py
@@ -1,19 +1,29 @@
 #!/usr/bin/env python3
-"""Main routes for the A.R.M ui"""
+"""
+Main routes for the A.R.M ui
+Covers
+- index [GET]
+- send_movies [GET]
+- error [GET]
+- errorhandler [GET]
+- setup [GET] -- penned for removal
+Other routes handled in flask blueprints
+- auth, database, history, jobs, logs, settings
+"""
 import os
 import json
 from pathlib import Path, PurePath
 from werkzeug.exceptions import HTTPException
-from werkzeug.routing import ValidationError
 from flask import Flask, render_template, request, send_file, flash, \
     redirect, url_for  # noqa: F401
 from flask.logging import default_handler  # noqa: F401
 from flask_login import login_required, UserMixin  # noqa: F401
+
 import arm.ui.utils as ui_utils
 from arm.ui import app, db, constants, json_api
 from arm.models import models as models
 import arm.config.config as cfg
-from arm.ui.forms import TitleSearchForm, ChangeParamsForm, DBUpdate
+from arm.ui.forms import DBUpdate
 from arm.ui.settings.ServerUtil import ServerUtil
 
 # This attaches the armui_cfg globally to let the users use any bootswatch skin from cdn
@@ -22,280 +32,6 @@ armui_cfg = ui_utils.arm_db_cfg()
 # Page definitions
 page_support_databaseupdate = "support/databaseupdate.html"
 redirect_settings = "/settings"
-
-
-@app.route('/error')
-def was_error(error):
-    """
-    Catch all error page
-    :return: Error page
-    """
-    return render_template(constants.ERROR_PAGE, title='error', error=error)
-
-
-@app.route('/setup')
-def setup():
-    """
-    This is the initial setup page for fresh installs
-    This is no longer recommended for upgrades
-
-    This function will do various checks to make sure everything can be setup for ARM
-    Directory ups, create the db, etc
-    """
-    perm_file = Path(PurePath(cfg.arm_config['INSTALLPATH'], "installed"))
-    app.logger.debug("perm " + str(perm_file))
-    # Check for install file and that db is correctly setup
-    if perm_file.exists() and ui_utils.setup_database(cfg.arm_config['INSTALLPATH']):
-        flash(f"{perm_file} exists, setup cannot continue. To re-install please delete this file.", "danger")
-        return redirect("/")
-    dir0 = Path(PurePath(cfg.arm_config['DBFILE']).parent)
-    dir1 = Path(cfg.arm_config['RAW_PATH'])
-    dir2 = Path(cfg.arm_config['TRANSCODE_PATH'])
-    dir3 = Path(cfg.arm_config['COMPLETED_PATH'])
-    dir4 = Path(cfg.arm_config['LOGPATH'])
-    arm_directories = [dir0, dir1, dir2, dir3, dir4]
-
-    try:
-        for arm_dir in arm_directories:
-            if not Path.exists(arm_dir):
-                os.makedirs(arm_dir)
-                flash(f"{arm_dir} was created successfully.", "success")
-    except FileNotFoundError as error:
-        flash(f"Creation of the directory {dir0} failed {error}", "danger")
-        app.logger.debug(f"Creation of the directory failed - {error}")
-    else:
-        flash("Successfully created all of the ARM directories", "success")
-        app.logger.debug("Successfully created all of the ARM directories")
-
-    try:
-        if ui_utils.setup_database(cfg['INSTALLPATH']):
-            flash("Setup of the database was successful.", "success")
-            app.logger.debug("Setup of the database was successful.")
-            perm_file = Path(PurePath(cfg.arm_config['INSTALLPATH'], "installed"))
-            write_permission_file = open(perm_file, "w")
-            write_permission_file.write("boop!")
-            write_permission_file.close()
-            return redirect(constants.HOME_PAGE)
-        flash("Couldn't setup database", "danger")
-        app.logger.debug("Couldn't setup database")
-        return redirect("/error")
-    except Exception as error:
-        flash(str(error))
-        app.logger.debug("Setup - " + str(error))
-        return redirect(constants.HOME_PAGE)
-
-
-@app.route('/json', methods=['GET'])
-@login_required
-def feed_json():
-    """
-    json mini API
-    This is used for all api/ajax calls this makes thing easier to read/code for
-    Adding a new function to the api is as simple as adding a new elif where GET[mode]
-    is your call
-    You can then add a function inside utils to deal with the request
-    """
-    mode = str(request.args.get('mode'))
-    return_json = {'mode': mode, 'success': False}
-    # Hold valid data (post/get data) we might receive from pages - not in here ? it's going to throw a key error
-    valid_data = {
-        'j_id': request.args.get('job'),
-        'searchq': request.args.get('q'),
-        'logpath': cfg.arm_config['LOGPATH'],
-        'fail': 'fail',
-        'success': 'success',
-        'joblist': 'joblist',
-        'mode': mode,
-        'config_id': request.args.get('config_id'),
-        'notify_id': request.args.get('notify_id')
-    }
-    # Valid modes that should trigger functions
-    valid_modes = {
-        'delete': {'funct': json_api.delete_job, 'args': ('j_id', 'mode')},
-        'abandon': {'funct': json_api.abandon_job, 'args': ('j_id',)},
-        'full': {'funct': json_api.generate_log, 'args': ('logpath', 'j_id')},
-        'search': {'funct': json_api.search, 'args': ('searchq',)},
-        'getfailed': {'funct': json_api.get_x_jobs, 'args': ('fail',)},
-        'getsuccessful': {'funct': json_api.get_x_jobs, 'args': ('success',)},
-        'fixperms': {'funct': ui_utils.fix_permissions, 'args': ('j_id',)},
-        'joblist': {'funct': json_api.get_x_jobs, 'args': ('joblist',)},
-        'send_item': {'funct': ui_utils.send_to_remote_db, 'args': ('j_id',)},
-        'change_job_params': {'funct': json_api.change_job_params, 'args': ('config_id',)},
-        'read_notification': {'funct': json_api.read_notification, 'args': ('notify_id',)}
-    }
-    if mode in valid_modes:
-        args = [valid_data[x] for x in valid_modes[mode]['args']]
-        return_json = valid_modes[mode]['funct'](*args)
-    return_json['notes'] = json_api.get_notifications()
-    return app.response_class(response=json.dumps(return_json, indent=4, sort_keys=True),
-                              status=200,
-                              mimetype=constants.JSON_TYPE)
-
-
-@app.route('/activerips')
-@login_required
-def rips():
-    """
-    This no longer works properly because of the 'transcoding' status
-    """
-    return render_template('activerips.html', jobs=models.Job.query.filter_by(status="active"))
-
-
-@app.route('/jobdetail')
-@login_required
-def jobdetail():
-    """
-    Page for showing in-depth details about a job
-
-    Shows Job/Config/Track class details
-    displays them in a clear and easy to ready format
-    """
-    job_id = request.args.get('job_id')
-    job = models.Job.query.get(job_id)
-    tracks = job.tracks.all()
-    search_results = ui_utils.metadata_selector("get_details", job.title, job.year, job.imdb_id)
-    if search_results and 'Error' not in search_results:
-        job.plot = search_results['Plot'] if 'Plot' in search_results else "There was a problem getting the plot"
-        job.background = search_results['background_url'] if 'background_url' in search_results else None
-    return render_template('jobdetail.html', jobs=job, tracks=tracks, s=search_results)
-
-
-@app.route('/titlesearch')
-@login_required
-def title_search():
-    """
-    The initial search page
-    """
-    job_id = request.args.get('job_id')
-    job = models.Job.query.get(job_id)
-    form = TitleSearchForm(request.args, meta={'csrf': False})
-    if form.validate():
-        flash(f'Search for {request.args.get("title")}, year={request.args.get("year")}', 'success')
-        return redirect(url_for('list_titles', title=request.args.get("title"),
-                                year=request.args.get("year"), job_id=job_id))
-    return render_template('titlesearch.html', title='Update Title', form=form, job=job)
-
-
-@app.route('/changeparams')
-@login_required
-def changeparams():
-    """
-    For updating Config params or changing/correcting job.disctype manually
-    """
-    config_id = request.args.get('config_id')
-    job = models.Job.query.get(config_id)
-    config = job.config
-    form = ChangeParamsForm(obj=config)
-    return render_template('changeparams.html', title='Change Parameters', form=form, config=config)
-
-
-@app.route('/customTitle')
-@login_required
-def customtitle():
-    """
-    For setting custom title for series with multiple discs
-    """
-    job_id = request.args.get('job_id')
-    ui_utils.job_id_validator(job_id)
-    job = models.Job.query.get(job_id)
-    form = TitleSearchForm(obj=job)
-    if request.args.get("title"):
-        args = {
-            'title': request.args.get("title"),
-            'title_manual': request.args.get("title"),
-            'year': request.args.get("year")
-        }
-        notification = models.Notifications(f"Job: {job.job_id} was updated",
-                                            f'Title: {job.title} ({job.year}) was updated to '
-                                            f'{request.args.get("title")} ({request.args.get("year")})')
-        db.session.add(notification)
-        ui_utils.database_updater(args, job)
-        flash(f'Custom title changed. Title={job.title}, Year={job.year}.', "success")
-        return redirect(url_for('home'))
-    return render_template('customTitle.html', title='Change Title', form=form, job=job)
-
-
-@app.route('/list_titles')
-@login_required
-def list_titles():
-    """
-    The search results page
-
-    This will display the returned search results from OMDB or TMDB from the users input search
-    """
-    title = request.args.get('title').strip() if request.args.get('title') else ''
-    year = request.args.get('year').strip() if request.args.get('year') else ''
-    job_id = request.args.get('job_id').strip() if request.args.get('job_id') else ''
-    if job_id == "":
-        app.logger.debug("list_titles - no job supplied")
-        flash(constants.NO_JOB, "danger")
-        raise ValidationError
-    job = models.Job.query.get(job_id)
-    form = TitleSearchForm(obj=job)
-    search_results = ui_utils.metadata_selector("search", title, year)
-    if search_results is None or 'Error' in search_results or (
-            'Search' in search_results and len(search_results['Search']) < 1):
-        app.logger.debug("No results found. Trying without year")
-        flash(f"No search results found for {title} ({year})<br/> Trying without year", 'danger')
-        search_results = ui_utils.metadata_selector("search", title, "")
-
-    if search_results is None or 'Error' in search_results or (
-            'Search' in search_results and len(search_results['Search']) < 1):
-        flash(f"No search results found for {title}", 'danger')
-    return render_template('list_titles.html', results=search_results, job_id=job_id,
-                           form=form, title=title, year=year)
-
-
-@app.route('/gettitle')
-@app.route('/select_title')
-@login_required
-def gettitle():
-    """
-    Used to display plot info from the search result page when the user clicks the title
-    and to forward the user to save the selected details
-
-    This was also used previously for the getdetails page but it no longer needed there
-    """
-    imdb_id = request.args.get('imdbID').strip() if request.args.get('imdbID') else None
-    job_id = request.args.get('job_id').strip() if request.args.get('job_id') else None
-    if imdb_id == "" or imdb_id is None:
-        app.logger.debug("gettitle - no imdb supplied")
-        flash("No imdb supplied", "danger")
-        raise ValidationError("No imdb supplied")
-    if job_id == "" or job_id is None:
-        app.logger.debug("gettitle - no job supplied")
-        flash(constants.NO_JOB, "danger")
-        raise ValidationError(constants.NO_JOB)
-    dvd_info = ui_utils.metadata_selector("get_details", None, None, imdb_id)
-    return render_template('showtitle.html', results=dvd_info, job_id=job_id)
-
-
-@app.route('/updatetitle')
-@login_required
-def updatetitle():
-    """
-    used to save the details from the search
-    """
-    # updatetitle?title=Home&amp;year=2015&amp;imdbID=tt2224026&amp;type=movie&amp;
-    #  poster=http://image.tmdb.org/t/p/original/usFenYnk6mr8C62dB1MoAfSWMGR.jpg&amp;job_id=109
-    job_id = request.args.get('job_id')
-    job = models.Job.query.get(job_id)
-    old_title = job.title
-    old_year = job.year
-    job.title = job.title_manual = ui_utils.clean_for_filename(request.args.get('title'))
-    job.year = job.year_manual = request.args.get('year')
-    job.video_type = job.video_type_manual = request.args.get('type')
-    job.imdb_id = job.imdb_id_manual = request.args.get('imdbID')
-    job.poster_url = job.poster_url_manual = request.args.get('poster')
-    job.hasnicetitle = True
-    notification = models.Notifications(f"Job: {job.job_id} was updated",
-                                        f'Title: {old_title} ({old_year}) was updated to '
-                                        f'{request.args.get("title")} ({request.args.get("year")})')
-    db.session.add(notification)
-    db.session.commit()
-    flash(f'Title: {old_title} ({old_year}) was updated to '
-          f'{request.args.get("title")} ({request.args.get("year")})', "success")
-    return redirect("/")
 
 
 @app.route('/')
@@ -355,6 +91,15 @@ def send_movies():
     return render_template('send_movies.html', job_list=return_job_list)
 
 
+@app.route('/error')
+def was_error(error):
+    """
+    Catch all error page
+    :return: Error page
+    """
+    return render_template(constants.ERROR_PAGE, title='error', error=error)
+
+
 @app.errorhandler(Exception)
 def handle_exception(sent_error):
     """
@@ -380,21 +125,53 @@ def handle_exception(sent_error):
     return render_template(constants.ERROR_PAGE, error=sent_error), 500
 
 
-@app.route('/update_arm', methods=['POST'])
-@login_required
-def update_git():
-    """Update arm via git command line"""
-    return ui_utils.git_get_updates()
-
-
-@app.route('/driveeject/<id>')
-@login_required
-def drive_eject(id):
+@app.route('/setup')
+def setup():
     """
-    Server System  - change state of CD/DVD/BluRay drive - toggle eject
+    This is the initial setup page for fresh installs
+    This is no longer recommended for upgrades
+
+    This function will do various checks to make sure everything can be setup for ARM
+    Directory ups, create the db, etc
     """
-    global redirect_settings
-    drive = models.SystemDrives.query.filter_by(drive_id=id).first()
-    drive.open_close()
-    db.session.commit()
-    return redirect(redirect_settings)
+    perm_file = Path(PurePath(cfg.arm_config['INSTALLPATH'], "installed"))
+    app.logger.debug("perm " + str(perm_file))
+    # Check for install file and that db is correctly setup
+    if perm_file.exists() and ui_utils.setup_database(cfg.arm_config['INSTALLPATH']):
+        flash(f"{perm_file} exists, setup cannot continue. To re-install please delete this file.", "danger")
+        return redirect("/")
+    dir0 = Path(PurePath(cfg.arm_config['DBFILE']).parent)
+    dir1 = Path(cfg.arm_config['RAW_PATH'])
+    dir2 = Path(cfg.arm_config['TRANSCODE_PATH'])
+    dir3 = Path(cfg.arm_config['COMPLETED_PATH'])
+    dir4 = Path(cfg.arm_config['LOGPATH'])
+    arm_directories = [dir0, dir1, dir2, dir3, dir4]
+
+    try:
+        for arm_dir in arm_directories:
+            if not Path.exists(arm_dir):
+                os.makedirs(arm_dir)
+                flash(f"{arm_dir} was created successfully.", "success")
+    except FileNotFoundError as error:
+        flash(f"Creation of the directory {dir0} failed {error}", "danger")
+        app.logger.debug(f"Creation of the directory failed - {error}")
+    else:
+        flash("Successfully created all of the ARM directories", "success")
+        app.logger.debug("Successfully created all of the ARM directories")
+
+    try:
+        if ui_utils.setup_database(cfg['INSTALLPATH']):
+            flash("Setup of the database was successful.", "success")
+            app.logger.debug("Setup of the database was successful.")
+            perm_file = Path(PurePath(cfg.arm_config['INSTALLPATH'], "installed"))
+            write_permission_file = open(perm_file, "w")
+            write_permission_file.write("boop!")
+            write_permission_file.close()
+            return redirect(constants.HOME_PAGE)
+        flash("Couldn't setup database", "danger")
+        app.logger.debug("Couldn't setup database")
+        return redirect("/error")
+    except Exception as error:
+        flash(str(error))
+        app.logger.debug("Setup - " + str(error))
+        return redirect(constants.HOME_PAGE)
